requirements.txt
# FlowGuard — Dependencies
# Install: pip install -r requirements.txt

# Core
networkx>=3.1
pandas>=2.0
numpy>=1.24
scipy>=1.11
pyyaml>=6.0
tqdm>=4.65

# ML / Anomaly Detection
scikit-learn>=1.3
xgboost>=2.0

# Deep Learning (Autoencoder)
torch>=2.1

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Solidity Parsing (optional — Slither)
# slither-analyzer>=0.10.0    # uncomment if solc is available

# Reporting
jinja2>=3.1
reportlab>=4.0

# Utilities
jsonlines>=3.1
requests>=2.31
